Crate dokkoo

Structs

Date

A Mokk file’s date-time metadata

Page

Generated data regarding a Mokk file

Functions

compile

Compiles a Mokk file; renders, makes note of the Mokk file (when, or if, the need arises)

create_liquid_parser

Creates a Liquid parser

get_contexts

Returns a Liquid object with a Page’s Liquid contexts

get_page_object

Returns an object with a Page’s context

get_permalink

Returns an expanded permalink value, for when shorthand is used

render

Returns a String with a &str’s Mokk file rendered

render_layouts

Render the layout(s) of a post recursively (should a layout have a layout of its own)

render_markdown

Render Markdown as HTML

split_frontmatter

Returns a tuple with a Mokk file’s frontmatter and contents, in that order