Function dokkoo::get_permalink

pub fn get_permalink(permalink: &str) -> String

Returns an expanded permalink value, for when shorthand is used

Arguments

permalink - A string slice that represents the permalink value specified in the Mokk file

Shorthand

date → /{{ page.data.collection }}/{{ page.date.year }}/{{ page.date.month }}/{{ page.date.day }}/{{ page.data.title }}.html

pretty → /{{ page.data.collection }}/{{ page.date.year }}/{{ page.date.month }}/{{ page.date.day }}/{{ page.data.title }}/index.html

ordinal → /{{ page.data.collection }}/{{ page.date.year }}/{{ page.date.y_day }}/{{ page.data.title }}.html

weekdate → /{{ page.data.collection }}/{{ page.date.year }}/W{{ page.date.week }}/{{ page.date.short_day }}/{{ page.data.title }}.html

none → /{{ page.data.collection }}/{{ page.data.title }}.html