Struct dokkoo::Date

pub struct Date {

    pub year: String,
    pub short_year: String,
    pub month: String,
    pub i_month: String,
    pub short_month: String,
    pub long_month: String,
    pub day: String,
    pub i_day: String,
    pub y_day: String,
    pub w_year: String,
    pub week: String,
    pub w_day: String,
    pub short_day: String,
    pub long_day: String,
    pub hour: String,
    pub minute: String,
    pub second: String,
    pub rfc_3339: String,
    pub rfc_2822: String,

}

A Mokk fileโs date-time metadata

Fields
year: String

Year with four digits

short_year: String

Year without the century (00..99)

month: String

Month (01..12)

i_month: String

Month without leading zeros

short_month: String

Three-letter month abbreviation, e.g. โJanโ

long_month: String

Full month name, e.g. โJanuaryโ

day: String

Day of the month (01..31)

i_day: String

Day of the month without leading zeros

y_day: String

Ordinal day of the year, with leading zeros. (001..366)

w_year: String

Week year which may differ from the month year for up to three days at the start of January and end of December

week: String

Week number of the current year, starting with the first week having a majority of its days in January (01..53)

w_day: String

Day of the week, starting with Monday (1..7)

short_day: String

Three-letter weekday abbreviation, e.g. โSunโ

long_day: String

Weekday name, e.g. โSundayโ

hour: String

Hour of the day, 24-hour clock, zero-padded (00..23)

minute: String

Minute of the hour (00..59)

second: String

Second of the minute (00..59)

rfc_3339: String

A Mokk fileโs date-time metadata, formatted per the RFC 3339 standard

rfc_2822: String

A Mokk fileโs date-time metadata, formatted per the RFC 2822 standard

Implementations

impl Date

pub fn new( year: String, short_year: String, month: String, i_month: String, short_month: String, long_month: String, day: String, i_day: String, y_day: String, w_year: String, week: String, w_day: String, short_day: String, long_day: String, hour: String, minute: String, second: String, rfc_3339: String, rfc_2822: String ) -> Date

impl Date

pub fn value_to_date(value: Option<&Value>, locale: Locale) -> Date

Convert a serde_yaml::Value object into a Date object

Arguments

value - The serde_yaml::Value object to convert

locale - A chrono::Locale object

pub fn chrono_to_date(datetime: DateTime<Utc>, locale: Locale) -> Date

Convert a chrono::DateTime object into a Date object

Arguments

datetime - A chrono::DateTime<chrono::Utc> object

locale - A chrono::Locale object