Struct dokkoo::Page

pub struct Page {
    pub document: Document,
    pub directory: String,
    pub name: String,
    pub url: String,
    pub year: String,
    pub short_year: String,
    pub month: String,
    pub i_month: String,
    pub short_month: String,
    pub long_month: String,
    pub day: String,
    pub i_day: String,
    pub y_day: String,
    pub w_year: String,
    pub week: String,
    pub w_day: String,
    pub short_day: String,
    pub long_day: String,
    pub hour: String,
    pub minute: String,
    pub second: String,
}

Generated data regarding a Mokk File

Fields
document: Document

A Page's Document

directory: String

Path to the File, not including the File itself

name: String

The File's base filename

url: Stringyear: String

Year with four digits

short_year: String

Year without the century (00..99)

month: String

Month (01..12)

i_month: String

Month without leading zeros

short_month: String

Three-letter month abbreviation, e.g. “Jan”

long_month: String

Full month name, e.g. “January”

day: String

Day of the month (01..31)

i_day: String

Day of the month without leading zeros

y_day: String

Ordinal day of the year, with leading zeros. (001..366)

w_year: String

Week year which may differ from the month year for up to three days at the start of January and end of December

week: String

Week number of the current year, starting with the first week having a majority of its days in January (01..53)

w_day: String

Day of the week, starting with Monday (1..7)

short_day: String

Three-letter weekday abbreviation, e.g. “Sun”

long_day: String

Weekday name, e.g. “Sunday”

hour: String

Hour of the day, 24-hour clock, zero-padded (00..23)

minute: String

Minute of the hour (00..59)

second: String

Second of the minute (00..59)