Crate dokkoo

Structs

Date
A File’s date-time metadata

Page
Generated data regarding a Mokk File

Functions

compile
Compiles a Mokk File; renders, makes note of the File (when, or if, the need arises)

create_liquid_parser
Creates a Liquid parser

get_contexts
Returns a Liquid object with a Page’s Liquid contexts

get_page_object
Returns an object with a Page’s context

get_permalink
Returns an expanded permalink value, for when shorthand is used

render
Returns a String with a &str’s File rendered

render_layouts
Render the layout(s) of a post recursively (should a layout have a layout of its own)

render_markdown
Render Markdown as HTML

split_frontmatter
Returns a tuple with a File’s frontmatter and contents, in that order