Crate dokkoo

Structs

Build

Data held in memory during the build process

Date

A Mokk file’s date-time metadata

Global

Build configuration data held in memory during the build process, from the global file

Page

Generated data regarding a Mokk file

Functions

create_liquid_parser

Creates a Liquid parser

default_locale

Gets the system locale, if available. Otherwise, defaults to en_US

default_locale_string

Gets a string representing the system locale, if available. Otherwise, defaults to ‘en_US’

get_global_context

Get the global context

get_permalink

Returns an expanded permalink value, for when shorthand is used

locale_string_to_locale

Gets a chrono::Locale object from a string

render_markdown

Render Markdown as HTML

split_frontmatter

Returns a tuple with a Mokk file’s frontmatter and contents, in that order